requirements.txt
streamlit==1.31.1
boto3==1.34.34
python-dotenv==1.0.0
pandas==2.2.0